+++ tests/testthat/test-gam.R
@@ -81,4 +81,14 @@ if (require("testthat") && require("insight") && require("mgcv")) {
   test_that("is_multivariate", {
     expect_false(is_multivariate(m1))
   })
+
+  test_that("find_variables", {
+    expect_equal(
+      find_variables(m1),
+      list(
+        response = "y",
+        conditional = c("s(x0)", "s(x1)", "s(x2)", "s(x3)")
+      )
+    )
+  })
 }