+++ tests/testthat/test-lme.R
@@ -2,29 +2,45 @@ if (require("testthat") && require("insight") && require("nlme") && require("lme
   context("insight, model_info")
 
   data("sleepstudy")
+  data(Orthodont)
   m1 <- lme(
     Reaction ~ Days,
     random = ~ 1 + Days | Subject,
     data = sleepstudy
   )
 
+  m2 <- lme(distance ~ age + Sex, data = Orthodont, random = ~ 1)
+
   test_that("model_info", {
     expect_true(model_info(m1)$is_linear)
   })
 
   test_that("find_predictors", {
-    expect_identical(find_predictors(m1), list(conditional = c("Days")))
+    expect_identical(find_predictors(m1), list(conditional = "Days"))
+    expect_identical(find_predictors(m2), list(conditional = c("age", "Sex")))
     expect_identical(find_predictors(m1, effects = "all"), list(conditional = "Days", random = "Subject"))
+    expect_identical(find_predictors(m2, effects = "all"), list(conditional = c("age", "Sex")))
     expect_identical(find_predictors(m1, flatten = TRUE), "Days")
     expect_identical(find_predictors(m1, effects = "random"), list(random = "Subject"))
   })
 
   test_that("find_response", {
     expect_identical(find_response(m1), "Reaction")
+    expect_identical(find_response(m2), "distance")
   })
 
   test_that("get_response", {
     expect_equal(get_response(m1), sleepstudy$Reaction)
+  })
+
+  test_that("find_random", {
+    expect_equal(find_random(m1), list(random = "Subject"))
+    expect_null(find_random(m2))
+  })
+
+  test_that("get_random", {
+    expect_equal(get_random(m1), data.frame(Subject = sleepstudy$Subject))
+    expect_warning(get_random(m2))
   })
 
   test_that("link_inverse", {
@@ -34,6 +50,7 @@ if (require("testthat") && require("insight") && require("nlme") && require("lme
   test_that("get_data", {
     expect_equal(nrow(get_data(m1)), 180)
     expect_equal(colnames(get_data(m1)), c("Reaction", "Days", "Subject"))
+    expect_equal(colnames(get_data(m2)), c("distance", "age", "Subject", "Sex"))
   })
 
   test_that("find_formula", {
@@ -45,11 +62,20 @@ if (require("testthat") && require("insight") && require("nlme") && require("lme
         random = as.formula("~1 + Days | Subject")
       )
     )
+    expect_length(find_formula(m2), 2)
+    expect_equal(
+      find_formula(m2),
+      list(
+        conditional = as.formula("distance ~ age + Sex"),
+        random = as.formula("~1")
+      )
+    )
   })
 
   test_that("find_terms", {
     expect_equal(find_terms(m1), list(response = "Reaction", conditional = "Days", random = "Subject"))
     expect_equal(find_terms(m1, flatten = TRUE), c("Reaction", "Days", "Subject"))
+    expect_equal(find_terms(m2), list(response = "distance", conditional = c("age", "Sex")))
   })
 
   test_that("n_obs", {
@@ -70,5 +96,12 @@ if (require("testthat") && require("insight") && require("nlme") && require("lme
     )
     expect_equal(nrow(get_parameters(m1)), 2)
     expect_equal(get_parameters(m1)$parameter, c("(Intercept)", "Days"))
+    expect_equal(
+      find_parameters(m2),
+      list(
+        conditional = c("(Intercept)", "age", "Sex"),
+        random = c("(Intercept)")
+      )
+    )
   })
 }